add component-structure
