+++ components/math/addition/add.js
@@ -1,6 +1,11 @@
-function add(a, b) {
-  sum = a + b;
-  console.log(sum);
-}
+import { getValues } from '../../getValues/getValues.js';
 
-//muss noch angepasst werden als iteration
+export function add(iterator) {
+  console.log('initialize addition');
+  getValues(iterator);
+  const addValue = getValues(iterator);
+  console.log(addValue);
+  const sum = parseInt(addValue.a, 10) + parseInt(addValue.b, 10);
+  console.log('sum', sum);
+  // const sum= a+b;
+}
