+++ components/math/multiplication/multiplication.js
@@ -1,0 +1,12 @@
+import { getValues } from '../../getValues/getValues.js';
+
+export function multiply(iterator) {
+    console.log('initialize addition');
+    getValues(iterator);
+    const addValue = getValues(iterator);
+    console.log(addValue);
+    const sum = parseInt(addValue.a, 10) / parseInt(addValue.b, 10);
+    console.log('sum', sum);
+    // const sum= a+b;
+  }
+
