+++ index.js
@@ -1,83 +1,24 @@
+import { getOperatorArray } from './components/getOperatorArray/getOperatorArray.js';
 import { analyzeArray } from './components/analyze/analyze.js';
 console.clear();
 import { handleOperators } from './components/handleOperators/handleOperators.js';
 import { operatorIndexArray } from './components/handleOperators/handleOperators.js';
+import { executeMath } from './components/executeMath/executeMath.js';
+
 //funktioniert , solange hinter der letzten ziffer ein plus ist ?!?! gelöst durch zusätzliche variable, die operator.length + 1 speichert
 export const string = '1-2*3+4-3/5';
 export const operatorIndexInString = [];
 export const operator = [];
 export let numbers = [];
-for (let i = 0; i < string.length; i++) {
-  if (
-    string[i] === '+' ||
-    string[i] === '-' ||
-    string[i] === '*' ||
-    string[i] === '/'
-  ) {
-    operatorIndexInString[operatorIndexInString.length] = i;
-    operator[operator.length] = string[i];
-  }
-}
-analyzeArray();
-handleOperators();
-// console.clear()
+getOperatorArray();
 console.log('numbers', numbers);
 console.log('operators', operatorIndexArray);
+analyzeArray();
+handleOperators();
+executeMath();
+
 console.log(operator);
-for (let i = 0; i < operator.length; i++) {
-  console.log('i', i);
-  console.log(operator[i]);
-  if (operator[i] === '/') {
-    console.log('DIVIDE', operator[i]);
-    divide(i);
-  }
-  if (operator[i] === '*') {
-    console.log('MULTIPLY', operator[i]);
-    multiply(i);
-  }
-  if (operator[i] === '+') {
-    console.log('ADD', operator[i]);
-    add(i);
-  }
-  if (operator[i] === '-') {
-    console.log('SUBSTRACT', operator[i]);
-    substract(i);
-  }
-}
-
-function divide(iterator) {
-  console.log('initialize division');
-  getValues(iterator);
-  // const sum = a/b;
-}
-function multiply(iterator) {
-  console.log('initialize multiplication');
-  getValues(iterator);
-  // const sum= a*b;
-}
-function substract(iterator) {
-  console.log('initialize substraction');
-  getValues(iterator);
-  // const sum= a-b;
-}
-function add(iterator) {
-  console.log('initialize addition');
-  getValues(iterator);
-  const addValue = getValues(iterator);
-  console.log(addValue);
-  const sum = parseInt(addValue.a, 10) + parseInt(addValue.b, 10);
-  console.log('sum', sum);
-  // const sum= a+b;
-}
-
-function getValues(iterator) {
-  const a = numbers[iterator];
-  const b = numbers[iterator + 1];
-  return {
-    a: a,
-    b: b
-  };
-}
+console.log('EOL');
 
 // muss noch funktion rein, wass passiert, wenn numbers[0] keine zahl ist, sondern beispielsweise ein minus!
 // vielleicht kann ich die substraktion in der form schreiben, dass sie das minus einfach auf die numbers[i] zieht
@@ -98,8 +39,6 @@ function getValues(iterator) {
 // console.log('b', b);
 // }
 // }
-
-console.log('EOL');
 
 // for (let i=0; i<number.length; i++) {
 //   console.log(i);
@@ -126,12 +65,21 @@ console.log('EOL');
 // logik: nehme zahl vor + und addiere sie zur zahl nach +
 // das ergebnis kommt in die variable zwischenergebnis[i]
 
-//Samstag 2023-02-04:
-//order in der if Abfrage, da das erste if höchste Priorität hat, die mathematischen gleichungen nach oben
-//die höher in der rangliste sind
-//also punkt vor strich
-//zuerst multiplikation and division
-//dann addition und substraktion
-//also nun erstmal die if anfrage schreiben für die operators
-//außerdem: der funktion für die aufspaltung von zahlen in strings einer variablen zuweisen,
-//um sie vor der if anfrage nach den operatoren direkt abzurufen
+// Samstag 2023-02-04:
+// order in der if Abfrage, da das erste if höchste Priorität hat, die mathematischen gleichungen nach oben
+// die höher in der rangliste sind
+// also punkt vor strich
+// zuerst multiplikation and division
+// dann addition und substraktion
+// also nun erstmal die if anfrage schreiben für die operators
+// außerdem: der funktion für die aufspaltung von zahlen in strings einer variablen zuweisen,
+// um sie vor der if anfrage nach den operatoren direkt abzurufen
+
+// 2023-03-08 01:27am
+// jsstructure implemented. index.js nearly clean now. import/export structure of components is working
+// next step is to get the math operations divide, multiply and substraction working as add.js
+// meaning: look for string before and after operator, then combing them with the assigend operator
+// after that i can think of implementing bracelets and then maybe have to think of ways to sort array
+// in order to execute the maths in correct order
+// also gotta think of combining all that what has been calculated to a sum then
+// but i need to take a nap. next day i got to write protocol for bootcamp
